+++ src/aind_dynamic_foraging_models/generative_model/__init__.py
@@ -1,1 +1,3 @@
 """Package for generative models of dynamic foraging behavior"""
+
+from .agent_q_learning import ForagerSimpleQ  # noqa: F401
