+++ src/aind_dynamic_foraging_models/generative_model/agent_q_learning.py
@@ -7,7 +7,7 @@ from typing import Literal
 import numpy as np
 
 from .act_functions import act_epsilon_greedy, act_softmax
-from .agent_q_learning_params import generate_pydantic_q_learning_params
+from .params.agent_q_learning_params import generate_pydantic_q_learning_params
 from .base import DynamicForagingAgentMLEBase
 from .learn_functions import learn_choice_kernel, learn_RWlike
