+++ tests/test_params_simpleQ.py
@@ -4,7 +4,7 @@ import unittest
 
 from pydantic import ValidationError
 
-from aind_dynamic_foraging_models.generative_model.agent_q_learning_params import (
+from aind_dynamic_foraging_models.generative_model.params.agent_q_learning_params import (
     generate_pydantic_q_learning_params,
 )
